Added a simple plan without post session save ability
Added a simple plan without post-session saving (later patch) with a simple improvement method. User input is yet to be implemented, however, it should be simple to replace the input when the var completedX is declared.

newELO had variable names changed for clarity.
In exercise.js the low weight was changed from (bodyWeight + 2) to (bodyWeight). Adding a bodyweight option to both weight and reps would be a good future feature.
home.html updated to include the new plans.js.