plans with max sets and dynamic goals
When the user had low max reps they would be asked to do an unreasonable number of sets. This could lead to overtraining for someone so new to the exercise.
Now a user has a max of 5sets and will be training to improve the number of reps per set. The adjustment function has been tailored to this new system to avoid slow improvement.

The program will now also work post 100reps. Asking the user to push themselves on the single set.

Progress can now be saved and loaded with planDaly100Save(pressUps, sitUps, squats) and loadLocalStorage() respectively.

the save function will not keep a detailed history yet.